refactor: Split scan module (#1120)
## Which issue does this PR close?



- Closes #.

## What changes are included in this PR?
Split Scan module, i was going through it and all the information for
scan was on one page (too long)
